feat: Migrated data class into new schemas.py
